version 2.0

+++ Controllers/CategoriasController.php
@@ -8,6 +8,7 @@ class CategoriasController {
 
     private $model;
     private $view;
+    private $user;
 
 	function __construct(){
 
@@ -15,6 +16,7 @@ class CategoriasController {
         $authHelper = new AuthHelper();
         $authHelper->checkLoggedIn();
         
+        $this->user = $authHelper->getLoggedUser();
         $this->model = new CategoriasModel();
         $this->view = new CategoriasView();
     }
@@ -25,28 +27,53 @@ class CategoriasController {
     }
 
     public function addCategoria(){
-        $this->model->addCategoria($_POST['tipo'],$_POST['descripcion']);
+        $tipo = strtoupper($_POST['tipo']);
+        $descripcion =  strtoupper($_POST['descripcion']);
+
+        if ($this->user->administrador == 1) {
+            $this->model->addCategoria($tipo, $descripcion);
+        }
+
         header("Location: " . CATEGORIAS);
     }
 
     public function displayCategoria($params = null) {
         $idCategoria = $params[':ID'];
-        $categoria = $this->model->get($idCategoria);
-        if ($categoria) // si existe la categoria
+        $categoria = $this->model->getCategoria($idCategoria);
+
+        if ($categoria)
             $this->view->displayCategoria($categoria);
         else
             $this->view->showError('La categoria no existe');
     }
 
-    public function editarCategoria() {
-        $fila = $_POST['fila'];
-        $this->model->editarCategoria($_POST['id_categoria'][$fila], $_POST['tipo'][$fila], $_POST['descripcion'][$fila]);
-        header("Location: " . CATEGORIAS . "/" . $_POST['id_categoria'][$fila] );
+    public function editarCategoria($params = null) {
+        /*
+            Dado un id, creo las variables correspondientes a cada valor de una categoria 
+            y, en caso de que la variable sea un string, la paso a mayuscula.
+            Ejemplo: postre -> POSTRE.
+            Luego, si el usuario es administrador invoco a "editarCategoria" de model
+            y redirecciono a la página donde está la categoría que se editó
+        */
+
+        $id_categoria = $params[':ID'];
+        $tipo = strtoupper($_POST['tipo']);
+        $descripcion =  strtoupper($_POST['descripcion']);
+
+        if ($this->user->administrador == 1) { 
+            $this->model->editarCategoria($id_categoria, $tipo, $descripcion);
+        }
+
+        header("Location: " . CATEGORIAS . "/" . $id_categoria );
     }
 
     public function deleteCategoria($params = null){
         $idCategoria = $params[':ID'];
-        $this->model->deleteCategoria($idCategoria);
+
+        if ($this->user->administrador == 1) {
+            $this->model->deleteCategoria($idCategoria);
+        }
+
         header("Location: " . CATEGORIAS);
     }
 }
